+++ gulpTasks/compile.js
@@ -3,4 +3,4 @@
 const gulp = require('gulp');
 const shell = require('gulp-shell');
 
-gulp.task('compile', shell.task([`webpack --process --progress --colors`]));
+gulp.task('compile', shell.task([`"node_modules/.bin/webpack" --process --progress --colors`]));