+++ src/commons/interfaces/ICoveoObject.ts
@@ -1,3 +1,5 @@
-export interface ICoveoObject {
+import { IClonable, IConfigurable } from '../collections/Dictionary';
+
+export interface ICoveoObject<T> extends IClonable<T>, IConfigurable<T> {
   getId(): string;
 }