+++ src/commons/utils/RequestUtils.ts
@@ -35,6 +35,7 @@ export class RequestUtils {
 
   static post(uri: string, apiKey: string, data: any): Promise<request.RequestResponse> {
     Logger.verbose(`${chalk.bgGreen.bold('POST')} ${chalk.underline(uri)}`);
+    Logger.insane(`${chalk.bgYellow.bold('POST')} ${JsonUtils.stringify(data, 0)}`);
     return new Promise((resolve: (value?: any) => void, reject: (error: any) => void) => {
       request(
         uri,
@@ -61,6 +62,7 @@ export class RequestUtils {
 
   static put(uri: string, apiKey: string, data: any): Promise<request.RequestResponse> {
     Logger.verbose(`${chalk.bgYellow.bold('PUT')} ${chalk.underline(uri)}`);
+    Logger.insane(`${chalk.bgYellow.bold('PUT')} ${JsonUtils.stringify(data, 0)}`);
     return new Promise((resolve: (value?: any) => void, reject: (error: any) => void) => {
       request(
         uri,
@@ -103,7 +105,7 @@ export class RequestUtils {
           if (err) {
             reject(err);
           } else {
-            if (response.statusCode === RequestUtils.NO_CONTENT) {
+            if (response.statusCode === RequestUtils.NO_CONTENT || response.statusCode === RequestUtils.OK) {
               resolve(response);
             } else {
               reject(JsonUtils.stringify(response.body));